+++ internal/sdk/common/types.go
@@ -292,9 +292,9 @@ type NestedSecurityGroupRule struct {
 type NestedSoftwareCatalog struct {
 	Catalog *NestedSoftwareCatalogCatalog `json:"catalog,omitempty" tfsdk:"catalog"`
 
-	EnabledCpuFamily map[string]interface{} `json:"enabled_cpu_family,omitempty" tfsdk:"enabled_cpu_family"`
+	EnabledCpuFamily *[]string `json:"enabled_cpu_family,omitempty" tfsdk:"enabled_cpu_family"`
 
-	EnabledCpuMicroarchitectures map[string]interface{} `json:"enabled_cpu_microarchitectures,omitempty" tfsdk:"enabled_cpu_microarchitectures"`
+	EnabledCpuMicroarchitectures *[]string `json:"enabled_cpu_microarchitectures,omitempty" tfsdk:"enabled_cpu_microarchitectures"`
 
 	PackageCount *int64 `json:"package_count,omitempty" tfsdk:"package_count"`
 
@@ -310,10 +310,6 @@ type NestedSoftwareCatalogCatalog struct {
 	Uuid *string `json:"uuid,omitempty" tfsdk:"uuid"`
 
 	Version *string `json:"version,omitempty" tfsdk:"version"`
-}
-type NestedSoftwareCatalogEnabledCpuFamily struct {
-}
-type NestedSoftwareCatalogEnabledCpuMicroarchitectures struct {
 }
 type NestedSoftwareCatalogPartition struct {
 	CpuArch *string `json:"cpu_arch,omitempty" tfsdk:"cpu_arch"`